[TASK-2605][PR-14] [+] Add TypeId<TContext> TypeId implicit conversions, remove redundant int cast

+++ MemoryAllocator/TypeIndexer/Data/TypeIdOf.cs
@@ -20,7 +20,7 @@ namespace Sapientia.TypeIndexer
 		static TypeIdOf()
 		{
 			IndexedTypes.GetContextTypeId(typeof(TContext), typeof(TType), out var rawId);
-			typeId = (int)rawId;
+			typeId = rawId;
 		}
 	}
 }
